requirements.txt
# API Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Vector Store
chromadb==0.5.23

# Local Embeddings (free, no API key needed)
sentence-transformers==3.3.1

# Local LLM via Ollama (free, no API key needed)
ollama==0.4.7

# Document Parsing
PyMuPDF==1.25.1

# Data Validation & Settings
pydantic==2.10.4
pydantic-settings==2.7.1

# Async HTTP
httpx==0.28.1

# Utilities
python-dotenv==1.0.1
structlog==24.4.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0

# OCR for scanned PDFs
pytesseract==0.3.13
Pillow==11.1.0